+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2.2.2
 
+* `Behaviour Builder`:
+  * Fixes in `CallbackQuery` triggers
+
 ## 2.2.1
 
 * `Versions`: